Add files via upload
This code automizes finding a star's coordinates by enabling the user to click on where the star is located and the code automatically saves the coordination and moves to the other image. The images are shown in a logarithmic scale so that the stars are more easily detectable. To facilitate easier and more precise selection there has been a zoom function added. This feature works by magnifying the area chosen by the cursor.